+++ Topic 1/src/exercise1to6/discount/CashDiscount.java
@@ -2,8 +2,8 @@ package exercise1to6.discount;
 
 import java.util.List;
 
-import exercise1to6.Item;
-import exercise1to6.Product;
+import exercise1to6.cart.Item;
+import exercise1to6.cart.Product;
 
 public class CashDiscount implements Discount {
