+++ Topic 1/src/exercise1to6/discount/Discount.java
@@ -2,7 +2,7 @@ package exercise1to6.discount;
 
 import java.util.List;
 
-import exercise1to6.Item;
+import exercise1to6.cart.Item;
 
 
 public interface Discount {
